--- client/src/lib/About/About.svelte
@@ -1,49 +1,0 @@
-<!--
- This file is Free Software under the Apache-2.0 License
- without warranty, see README.md and LICENSES/Apache-2.0.txt for details.
-
- SPDX-License-Identifier: Apache-2.0
-
- SPDX-FileCopyrightText: 2024 German Federal Office for Information Security (BSI) <https://www.bsi.bund.de>
- Software-Engineering: 2024 Intevation GmbH <https://intevation.de>
--->
-
-<script lang="ts">
-  import { A, P, Li, List } from "flowbite-svelte";
-  import { onMount } from "svelte";
-  import SectionHeader from "$lib/SectionHeader.svelte";
-  import { appStore } from "$lib/store";
-  import { request } from "$lib/utils";
-  import ErrorMessage from "$lib/Messages/ErrorMessage.svelte";
-
-  let version: string = "Retrieving Version from server";
-  let error: string;
-  onMount(async () => {
-    if ($appStore.app.keycloak.authenticated) {
-      const response = await request("api/about", "GET");
-      if (response.ok) {
-        const backendInfo = response.content;
-        version = backendInfo.version;
-      } else if (response.error) {
-        error = response.error;
-      }
-    }
-  });
-</script>
-
-<SectionHeader title="About ISDuBA"></SectionHeader>
-
-<P>
-  <A href="https://github.com/ISDuBA/" class="underline hover:no-underline"
-    >Visit the ISDuBA project on Github</A
-  ></P
->
-{#if $appStore.app.keycloak.authenticated}
-  <P class="mt-3">
-    Versions:
-    <List tag="ul" class="space-y-1" list="none">
-      <Li liClass="ml-3">ISDuBA: {version}</Li>
-    </List>
-    <ErrorMessage message={error} plain={true}></ErrorMessage>
-  </P>
-{/if}
