+++ client/src/lib/SideNav.svelte
@@ -12,30 +12,22 @@
   import { Heading, Sidebar, SidebarWrapper, SidebarGroup, SidebarItem } from "flowbite-svelte";
   import { appStore } from "$lib/store";
   import { page } from "$app/stores";
-  import { PUBLIC_KEYCLOAK_URL } from "$env/static/public";
-  async function logout() {
-    $appStore.app.keycloak.logout();
-  }
 
   let notactivated =
     "flex items-center p-2 text-base font-normal text-gray-400 dark:text-gray-400 hover:bg-primary-100 hover:text-primary-900";
 
-  function login() {
-    $appStore.app.keycloak.login();
-  }
   $: activeUrl = "/" + $page.url.hash;
   let activeClass =
     "flex items-center p-2 text-base font-normal text-primary-900 bg-primary-200 dark:bg-primary-700 dark:text-white hover:bg-primary-100 dark:hover:bg-gray-700";
   let nonActiveClass =
     "flex items-center p-2 text-base font-normal text-white dark:text-white hover:bg-primary-100 hover:text-primary-900";
-  let profileUrl = PUBLIC_KEYCLOAK_URL + "/realms/isduba/account/#/";
 </script>
 
-<Sidebar class="h-screen bg-primary-700 p-2" {activeUrl} {activeClass} {nonActiveClass}>
-  <SidebarWrapper class="bg-primary-700">
-    <Heading class="mb-6 text-white">ISDuBA</Heading>
-    <SidebarGroup class="bg-primary-700">
-      {#if $appStore.app.keycloak.authenticated}
+{#if $appStore.app.keycloak && ($appStore.app.keycloak.authenticated || $appStore.app.sessionExpired)}
+  <Sidebar class="bg-primary-700 h-screen p-2" {activeUrl} {activeClass} {nonActiveClass}>
+    <SidebarWrapper class="bg-primary-700">
+      <Heading class="mb-6 text-white">ISDuBA</Heading>
+      <SidebarGroup class="bg-primary-700">
         <!-- Entries which are available after login should go here-->
         <SidebarItem label="Home" href="/#/">
           <svelte:fragment slot="icon">
@@ -72,29 +64,17 @@
             <i class="bx bx-cog"></i>
           </svelte:fragment>
         </SidebarItem>
-        <SidebarItem label="Profile" target="_blank" href={profileUrl}>
-          <svelte:fragment slot="icon">
-            <i class="bx bx-user"></i>
-          </svelte:fragment>
-        </SidebarItem>
-        <SidebarItem
-          on:click={logout}
-          label="Logout ({$appStore.app.userProfile.firstName} {$appStore.app.userProfile
-            .lastName})"
-        >
-          <svelte:fragment slot="icon">
-            <i class="bx bx-power-off"></i>
-          </svelte:fragment>
-        </SidebarItem>
-      {:else}
-        <!-- Entries which should be available only if not logged in should go here-->
-        <SidebarItem on:click={login} label="Login">
-          <svelte:fragment slot="icon">
-            <i class="bx bx-log-in"></i>
-          </svelte:fragment>
-        </SidebarItem>
-      {/if}
-      <SidebarItem label="About" href="/#/about"></SidebarItem>
-    </SidebarGroup>
-  </SidebarWrapper>
-</Sidebar>
+        {#if !$appStore.app.sessionExpired}
+          <SidebarItem
+            label={$appStore.app.keycloak.idTokenParsed.preferred_username}
+            href="/#/login"
+          >
+            <svelte:fragment slot="icon">
+              <i class="bx bx-user"></i>
+            </svelte:fragment>
+          </SidebarItem>
+        {/if}
+      </SidebarGroup>
+    </SidebarWrapper>
+  </Sidebar>
+{/if}
